코드 정리 , 던전 층 올라갈때마다 몬스터 강해지도록 설정 (2층이상부터  층 /0.5 배수)

+++ Assets/Scripts/Unit/LivingEntity/Monster/Monster.cs
@@ -42,7 +42,7 @@ public class Monster : LivingEntity
     [SerializeField] protected int _attackPattern; public int attackPattern { get { return _attackPattern; } }
     [SerializeField] protected string _description; public string description { get { return _description; } }
     [SerializeField] protected float _attackTime; public float attackTime { get { return _attackTime; } }
-    [SerializeField] protected int _attackDamage; public int attackDamage { get { return _attackDamage; } set { _attackDamage = value; } }
+    [SerializeField] protected float _attackDamage; public float attackDamage { get { return _attackDamage; } set { _attackDamage = value; } }
 
     [SerializeField] protected Material _dissolveMaterial; public Material dissolveMaterial { get { return _dissolveMaterial; } }
     [SerializeField] protected Material _nonDissolveMaterial; public Material nonDissolveMaterial { get { return _nonDissolveMaterial; } }
